Refactor sso services
