+++ src/ares/service/src/apis/admin/controllers/sso.ts
@@ -2,7 +2,10 @@ import { Paginator } from '@lowerdeck/pagination';
 import { v } from '@lowerdeck/validation';
 import { env } from '../../../env';
 import { adminService } from '../../../services/admin';
-import { ssoService } from '../../../services/sso';
+import { ssoConnectionService } from '../../../services/sso/connection';
+import { ssoDirectoryService } from '../../../services/sso/directory';
+import { ssoSetupService } from '../../../services/sso/setup';
+import { ssoTenantService } from '../../../services/sso/tenant';
 import { adminApp } from '../middleware/admin';
 import {
   ssoConnectionPresenter,
@@ -22,7 +25,7 @@ export let ssoController = adminApp.controller({
     )
     .do(async ({ input }) => {
       let app = await adminService.getApp({ appId: input.appId });
-      let paginator = await ssoService.listTenants({ app });
+      let paginator = await ssoTenantService.listTenants({ app });
       let list = await paginator.run(input);
       return Paginator.presentLight(list, ssoTenantPresenter);
     }),
@@ -35,7 +38,7 @@ export let ssoController = adminApp.controller({
       })
     )
     .do(async ({ input }) => {
-      let tenant = await ssoService.getTenantById({ tenantId: input.id });
+      let tenant = await ssoTenantService.getTenantById({ tenantId: input.id });
       return ssoTenantPresenter(tenant);
     }),
 
@@ -49,7 +52,7 @@ export let ssoController = adminApp.controller({
     )
     .do(async ({ input }) => {
       let app = await adminService.getApp({ appId: input.appId });
-      let tenant = await ssoService.createTenant({
+      let tenant = await ssoTenantService.createTenant({
         app,
         input: { name: input.name }
       });
@@ -65,8 +68,8 @@ export let ssoController = adminApp.controller({
       })
     )
     .do(async ({ input }) => {
-      let tenant = await ssoService.getTenantById({ tenantId: input.tenantId });
-      let setup = await ssoService.createSetup({
+      let tenant = await ssoTenantService.getTenantById({ tenantId: input.tenantId });
+      let setup = await ssoSetupService.createSetup({
         tenant,
         input: {
           redirectUri: `${env.service.ARES_ADMIN_URL}/apps/${input.appId}?sso_setup_complete=1`
@@ -88,8 +91,8 @@ export let ssoController = adminApp.controller({
       )
     )
     .do(async ({ input }) => {
-      let tenant = await ssoService.getTenantById({ tenantId: input.tenantId });
-      let paginator = await ssoService.listConnections({ tenant });
+      let tenant = await ssoTenantService.getTenantById({ tenantId: input.tenantId });
+      let paginator = await ssoConnectionService.listConnections({ tenant });
       let list = await paginator.run(input);
       return Paginator.presentLight(list, ssoConnectionPresenter);
     }),
@@ -104,12 +107,12 @@ export let ssoController = adminApp.controller({
       })
     )
     .do(async ({ input }) => {
-      let tenant = await ssoService.getTenantById({ tenantId: input.tenantId });
-      let connection = await ssoService.getConnectionById({
+      let tenant = await ssoTenantService.getTenantById({ tenantId: input.tenantId });
+      let connection = await ssoConnectionService.getConnectionById({
         tenant,
         connectionId: input.connectionId
       });
-      let updated = await ssoService.setConnectionStatus({
+      let updated = await ssoConnectionService.setConnectionStatus({
         tenant,
         connection,
         status: input.status
@@ -129,12 +132,12 @@ export let ssoController = adminApp.controller({
       })
     )
     .do(async ({ input }) => {
-      let tenant = await ssoService.getTenantById({ tenantId: input.tenantId });
-      let connection = await ssoService.getConnectionById({
+      let tenant = await ssoTenantService.getTenantById({ tenantId: input.tenantId });
+      let connection = await ssoConnectionService.getConnectionById({
         tenant,
         connectionId: input.connectionId
       });
-      let { directory, scim } = await ssoService.createDirectory({
+      let { directory, scim } = await ssoDirectoryService.createDirectory({
         tenant,
         connection,
         input: {
@@ -161,12 +164,12 @@ export let ssoController = adminApp.controller({
       )
     )
     .do(async ({ input }) => {
-      let tenant = await ssoService.getTenantById({ tenantId: input.tenantId });
-      let connection = await ssoService.getConnectionById({
+      let tenant = await ssoTenantService.getTenantById({ tenantId: input.tenantId });
+      let connection = await ssoConnectionService.getConnectionById({
         tenant,
         connectionId: input.connectionId
       });
-      let paginator = await ssoService.listDirectories({ connection });
+      let paginator = await ssoDirectoryService.listDirectories({ connection });
       let list = await paginator.run(input);
       return Paginator.presentLight(list, ssoDirectoryPresenter);
     }),
@@ -181,12 +184,12 @@ export let ssoController = adminApp.controller({
       })
     )
     .do(async ({ input }) => {
-      let tenant = await ssoService.getTenantById({ tenantId: input.tenantId });
-      let connection = await ssoService.getConnectionById({
+      let tenant = await ssoTenantService.getTenantById({ tenantId: input.tenantId });
+      let connection = await ssoConnectionService.getConnectionById({
         tenant,
         connectionId: input.connectionId
       });
-      let directory = await ssoService.getDirectoryById({
+      let directory = await ssoDirectoryService.getDirectoryById({
         tenant,
         connection,
         directoryId: input.directoryId
@@ -205,17 +208,17 @@ export let ssoController = adminApp.controller({
       })
     )
     .do(async ({ input }) => {
-      let tenant = await ssoService.getTenantById({ tenantId: input.tenantId });
-      let connection = await ssoService.getConnectionById({
+      let tenant = await ssoTenantService.getTenantById({ tenantId: input.tenantId });
+      let connection = await ssoConnectionService.getConnectionById({
         tenant,
         connectionId: input.connectionId
       });
-      let directory = await ssoService.getDirectoryById({
+      let directory = await ssoDirectoryService.getDirectoryById({
         tenant,
         connection,
         directoryId: input.directoryId
       });
-      let updated = await ssoService.setDirectoryStatus({
+      let updated = await ssoDirectoryService.setDirectoryStatus({
         tenant,
         connection,
         directory,
@@ -228,7 +231,7 @@ export let ssoController = adminApp.controller({
     .handler()
     .input(Paginator.validate())
     .do(async ({ input }) => {
-      let paginator = await ssoService.listGlobalTenants();
+      let paginator = await ssoTenantService.listGlobalTenants();
       let list = await paginator.run(input);
       return Paginator.presentLight(list, t => ({
         ...ssoTenantPresenter(t),
@@ -245,8 +248,8 @@ export let ssoController = adminApp.controller({
       })
     )
     .do(async ({ input }) => {
-      let tenant = await ssoService.getTenantById({ tenantId: input.id });
-      let updated = await ssoService.setGlobal({ tenant, isGlobal: input.isGlobal });
+      let tenant = await ssoTenantService.getTenantById({ tenantId: input.id });
+      let updated = await ssoTenantService.setGlobal({ tenant, isGlobal: input.isGlobal });
       return ssoTenantPresenter(updated);
     })
 });
