+++ src/ares/service/src/services/sso/types.ts
@@ -1,0 +1,7 @@
+import type { SsoConnection, SsoDirectory, SsoTenant } from '../../../prisma/generated/client';
+
+export type SsoDirectoryWithApp = SsoDirectory & {
+  connection?: SsoConnection & {
+    tenant?: SsoTenant;
+  };
+};
